\name{estSNR}
\alias{estSNR}
\title{Generate a random sparse test signal s}
\usage{
  estSNR(estimated.s, true.s)
}
\arguments{
  \item{estimated.s}{estimated.s Calculated s vector}

  \item{true.s}{true.s If known, the true s vector.}
}
\value{
  SNR
}
\description{
  Generate a random sparse test signal s
}
\keyword{SNR}

